Moves backbone files
